+++ components/core/src/memfault_batched_events.c
@@ -11,12 +11,6 @@
 #include "memfault/core/sdk_assert.h"
 #include "memfault/util/cbor.h"
 
-
-static void prv_fill_header_cb(void *ctx, uint32_t offset, const void *buf, size_t buf_len) {
-  uint8_t *header_buf = (uint8_t *)ctx;
-  memcpy(&header_buf[offset], buf, buf_len);
-}
-
 void memfault_batched_events_build_header(
     size_t num_events, sMemfaultBatchedEventsHeader *header_out) {
   MEMFAULT_SDK_ASSERT(header_out != NULL);
@@ -28,7 +22,7 @@ void memfault_batched_events_build_header(
 
   // there's multiple events to read. We will add a header to indicate the total count
   sMemfaultCborEncoder encoder;
-  memfault_cbor_encoder_init(&encoder, prv_fill_header_cb, header_out->data,
+  memfault_cbor_encoder_init(&encoder, memfault_cbor_encoder_memcpy_write, header_out->data,
                              sizeof(header_out->data));
   memfault_cbor_encode_array_begin(&encoder, num_events);
   header_out->length = memfault_cbor_encoder_deinit(&encoder);